+++ bin/tasks/check-dependencies.sh
@@ -1,5 +1,23 @@
 #!/bin/bash
 
+# Check if block is registered with the same name
+if grep -Fxq ${BLOCK_NAME} ${PROJECT_THEME_PATH}/functions.php
+then
+  echo "${RED}Found exactly the same block name registered in functions.php... The script will now quit.${TXTRESET}" 1>&2
+  exit
+else
+  echo " "
+fi
+
+# Let's ensure we have acf-json folder
+mkdir -p ${PROJECT_THEME_PATH}/acf-json
+
+# Ensure file permissions for that folder
+source ${SCRIPTS_LOCATION}/tasks/permissions.sh
+
+# Create blocks directory if it does not exist
+mkdir -p ${PROJECT_THEME_PATH}/template-parts/blocks
+
 # Let's go to the project directory
 cd ${PROJECTS_HOME}/${PROJECT_NAME}